ASN.1 OID parsing (nw)
